@@ -59,8 +59,9 @@ public enum UIModel {
 				InetAddress host = null;
 				if (hostStr != null && !hostStr.isEmpty()) {
 					host = InetAddress.getByName(hostStr);
+				} else {
+					host = InetAddress.getByName("localhost");
 				}
-
 				SMTPServerHandler.INSTANCE.startServer(port, host);
 			} catch (NumberFormatException e) {
 				throw new InvalidPortException(e);
